// frontend/apps/pathways/app/jobs/_modules/types.d.ts
export type GetJobRoleResponse = {
  title: string;
  job_category: string;
  pathway: string;
  position_type: string;
  level_of_experience: string;
  role_description: string;
  education_level: string;
  special_educational_requirements: string;
  workplace_type: string;
  city: string;
  state: string;
  country: string;
  pay_type: string;
  minimum_pay: number;
  maximum_pay: number;
  pay_frequency: string;
  closing_date: string;
  external_link: string;
  required_files: string[];
  status: string;
  primary_responsibilities: string;
  required_qualifications: string;
  desired_qualifications: string;
  compensation_benefits: string;
  id: string;
  company_id: string;
  company_name: string;
  company_logo_url: string;
  is_bipoc_owned: boolean;
  created_at: string;
  user_id: string;
  stage: {
    [key: string]: boolean;
  };
};

export type Filters = {
  title?: string;
  job_type?: string[];
  position_type?: string[];
  level_of_experience?: string[];
  pathway?: string[];
  workplace_type?: string[];
  city?: string;
  state?: string;
  country?: string;
  pay_type?: string;
  minimum_pay?: number[];
  maximum_pay?: number[];
  pay_frequency?: string;
  company_name?: string;
  is_bipoc_owned?: boolean;
};